Changes for review.
Using the config in setup meant that a bunch of tests failed that
used the Consumer without properly initializing it.
They're now fixed, but the tests could probably do with cleaning
up at some point.

This included one ostensibly external test in consumer_test.go.
That test didn't seem to be sufficiently different from the others
to justify the extra file, so it's moved into the other test file now.