consumer: use default retry interval when MaxRetryInterval is zero
It seems that kafka-go is setting MaxRetryInterval to zero
if a service doesn't explicitly set a retry interval, and
that will result in backoff with no time limit on the retry interval,
so use the default retry interval whenever MaxRetryInterval
is zero.

As retries have been removed from felice in master but we want
to retrofix this without upgrading kafka-go to the latest felice,
this is targetted at the sendmessages-v0.4.0 feature branch,
which will also contain the new SendMessages feature backported.